testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/history_pushstate_err.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>history_pushState SECURITY_ERR</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		try {
			window.history.pushState(1, document.title, 'http://www.microsoft.com/test.html');
		} catch (e) {
			test(function() {
				assert_equals(e.code, e["SECURITY_ERR"], "SECURITY_ERR is not raised.");
			});
		}

	</script>
 </body>
</html>
